CHROMIUM: ARM: dts: rockchip: assigned parents for vop dclks for rk3288
The VOP DCLK is used to generate pixel clocks.

For veyron, we have decided to permanently assign vop0 (the big VOP) for
use with HDMI and vop1 (the little VOP) for use with eDP.

Furthermore, to allow us to generate a wide range of precise pixel clocks,
we will be dedicating the NPLL exclusively for use as the parent clock
for VOP0/HDMI.

To implement the exclusive assignment of NPLL in the kernel, we remove
the NPLL entry from all clock muxes that would otherwise be able to select
it (such as vop1).  For vop0, we remove all choices *except* NPLL.

Before booting the kernel, the bios will configure vop0 and vop1 as it
sees fit - potentially assigning NPLL to vop1 and some other PLL to vop0.

Thus, at boot it is possible that from the kernel's perspective, these
clocks are orphans.  To fix this, we explicitly assign their clock parents
to ensure that they are properly parented no matter what state they are
when the kernel boots.

Signed-off-by: Yakir Yang <ykk@rock-chips.com>
Signed-off-by: Daniel Kurtz <djkurtz@chromium.org>

BUG=chromium:941638, b:35542882
TEST=Check /sys/kernel/debug/clk/clk_summary

Change-Id: I384063d644437521670d5eeceee5867c387f58cf
Reviewed-on: https://chromium-review.googlesource.com/243160
Trybot-Ready: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>
Reviewed-by: Douglas Anderson <dianders@chromium.org>
Commit-Queue: Daniel Kurtz <djkurtz@chromium.org>
(cherry picked from commit 9e621622615f3eaf81d214d6f3f6b55e842a7df4)
[dianders: slightly adjusted the commit message from 3.14]
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1574580
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>