CHROMIUM: clk: rockchip: Register some debug clocks as global
In "CHROMIUM: soc: rockchip: Add panic handler for rk3288" we'd like
to access a few clocks without any associated device. That's pretty
ugly and gross, but since upstream is likely to have a very different
solution it doesn't seem to make sense to try to architect something
cleaner just for our tree.

We'll instead just register a few global clocks.  If this were common
we would move it to a common rockchip clock file, but since this is
just for something local to our tree we'll put it in the rk3288 file.

Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/292860
Reviewed-by: Andrew Bresticker <abrestic@chromium.org>

BUG=chromium:941638
TEST=see CL:I576edf75f23c4d0c5c1826e1b0a3017200e7c5ed

Change-Id: Ie83bdf976cdb39fb2123af58f42c9ceac957d320
Signed-off-by: Matthias Kaehlcke <mka@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1630713
Commit-Ready: Douglas Anderson <dianders@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Douglas Anderson <dianders@chromium.org>