CHROMIUM: ARM: Add a debugfs file for dumping CP14/CP15 regs
This is a squash of two CLs in the Chrome OS 3.14 kernel:

10e326f3dd09 CHROMIUM: ARM: Add a debugfs file for dumping CP14/CP15 regs
de1df81c53c8 CHROMIUM: ARM: Add internal feature register to CP14/CP15 debugfs

In order to confirm that certain errata have been applied properly, we'd
like to be able to read some CP14 / CP15 registers from userspace.
Normally that's not allowed, but for some registers there's no harm in
providing read-only access to userspace.

Let's add a bit of infrastructure here for doing this.  For now we'll
only add one register: one that's needed for confirming an errata has
been applied properly on A12 / A17.

BUG=chromium:941638
TEST=taskset -c $i cat /sys/kernel/debug/arm_coprocessor_debug
TEST=kernel_CheckArmErrata

Change-Id: Ib4685e505ae0411d2dd3e94a0785b5f470523f29
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1586210
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>
Reviewed-by: Guenter Roeck <groeck@chromium.org>