CHROMIUM: add vmstat counts for different page fault types
VM event counters are added to keep track of anonymous
vs. file vs. shmem page faults.  They are: pgmajfault_a,
pgmajfault_f and pgmajfault_s.  These are useful to
analyze system performance, particularly when the cost
of a fault for a file page is very different from that
of an anonymous page, as would happen, for instance, in
the presence of zram.

The PGMAJFAULT counter is no longer directly maintained,
for performance.  Instead the three new counters are added
whenever the total count is needed.

This adds entries to /proc/vmstat.  I considered adding them at
the end for better compatibility, but the contents of /proc/vmstat
are already highly variable depending on config options.  It is
still worth asking whether the new counters should NOT contain
"pgmajfault" as a substring of their name.

BUG=chromium:718270
TEST=checked that new fields are printed

Reviewed-on: https://chromium-review.googlesource.com/470508
Commit-Ready: Luigi Semenzato <semenzato@chromium.org>
Tested-by: Luigi Semenzato <semenzato@chromium.org>
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>
Reviewed-by: Justin TerAvest <teravest@chromium.org>
Signed-off-by: Luigi Semenzato <semenzato@chromium.org>

Change-Id: I639cdd543d9bd72996ac2c585f06ae9dd581acae
Reviewed-on: https://chromium-review.googlesource.com/1583122
Commit-Ready: Luigi Semenzato <semenzato@chromium.org>
Tested-by: Luigi Semenzato <semenzato@chromium.org>
Reviewed-by: Brian Geffon <bgeffon@chromium.org>
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>