UPSTREAM: media: v4l: Add support for V4L2_BUF_TYPE_META_OUTPUT
The V4L2_BUF_TYPE_META_OUTPUT mirrors the V4L2_BUF_TYPE_META_CAPTURE with
the exception that it is an OUTPUT type. The use case for this is to pass
buffers to the device that are not image data but metadata. The formats,
just as the metadata capture formats, are typically device specific and
highly structured.

Signed-off-by: Sakari Ailus <sakari.ailus@linux.intel.com>
Acked-by: Hans Verkuil <hans.verkuil@cisco.com>
Reviewed-by: Tomasz Figa <tfiga@chromium.org>
Tested-by: Tian Shu Qiu <tian.shu.qiu@intel.com>
Signed-off-by: Mauro Carvalho Chehab <mchehab+samsung@kernel.org>

BUG=b:135068037
TEST=build and boot to shell

Change-Id: Ieaa0836291f2a13f4633df86fc2fecacce0c3e52
Signed-off-by: Bibby Hsieh <bibby.hsieh@mediatek.com>
(cherry picked from commit 72148d1a57e7c76745e68c94ad5d235240d26ac8)
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/1662146
Reviewed-by: Nicolas Boichat <drinkcat@chromium.org>
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Tested-by: Nicolas Boichat <drinkcat@chromium.org>
Commit-Queue: Nicolas Boichat <drinkcat@chromium.org>