CHROMIUM: hid: quickstep: Initial add of driver
The ChomeOS Touch team uses a usb device built in-house to measure the
drag latency on touchpads.  This device communicates with the OS over
USB using a custom driver.  This CL adds that driver so that by default
Chromebooks will have the driver ready.  Previously, the kernel had to
be rebuilt for every device we wanted to test every time, so it will
save a lot of time on our end to already have it in there.

The device (named Quickstep) consists of a laser that lays across the
touchpad and hits a phototransistor on the other side.  The device
reports over USB HIDRAW a 1 or a 0 immediates each time the laser is
broken/unbroken by a finger inbetween the sensor and the laser.  This
driver simply listens for these events and records the time at which
they occur.  These events are stored in memory and can be accessed via
sysfs entries for later analysis by the latency-measuring scripts.

BUG=chromium:356342
TEST=built and used with x86-generic which uses the 3.10 kernel

Change-Id: I72b0cc8fe009c2a83ab4a8b03747749693fbf03a
Signed-off-by: Charlie Mooney <charliemooney@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/191514
Reviewed-by: Olof Johansson <olofj@chromium.org>

[benzh: 3.14 rebase. Resolved trivial conflicts]
Signed-off-by: Ben Zhang <benzh@chromium.org>

[dtor: 4.4 rebase, resolved trivial conflicts in drivers/hid/hid-core.c
 and squashed:
 - CHROMIUM: HID: hid-quickstep - Change get kernel time function
 - CHROMIUM: hid: quickstep: Remove world-writable access from "clear" device
]
Signed-off-by: Filipe Brandenburger <filbranden@chromium.org>
Signed-off-by: Dmitry Torokhov <dtor@chromium.org>
Conflicts:
	drivers/hid/hid-core.c
[rebase412(groeck): Context conflicts]
Signed-off-by: Guenter Roeck <groeck@chromium.org>
[rebase414(lannm): Trivial conflicts]
Signed-off-by: Lann Martin <lannm@chromium.org>
Conflicts:
	drivers/hid/Makefile
	drivers/hid/hid-core.c
	drivers/hid/hid-ids.h
[rebase419(groeck): Context conflicts]
Signed-off-by: Guenter Roeck <groeck@chromium.org>