CHROMIUM: Revert "Bluetooth: btusb: driver to enable the usb-wakeup feature"
This reverts commit a0085f2510e8 ("Bluetooth: btusb: driver to
enable the usb-wakeup feature"). The reason is that this causes any inband
communication from the btusb device to cause the system to wake up. We
know that the BT controller could be scanning when we decide to suspend
and will send in the results once the scan is complete, thus failing the
suspend entry. This is a temporary work around until the bug 67489211
is solved. We can put this back once the bluetoothd starts to listen to
powerd for suspend requests and do the right thing.

(Reverting this commit has the side effect that BLE mouse clicks do not
wake up the system from suspend, but that is lesser price to pay than no
going to suspend).

Signed-off-by: Rajat Jain <rajatja@google.com>
BUG=b:67489211, b:37256085, b:134132353, b:128319712
TEST=Suspend resume via a BLuetooth mouse

Change-Id: I6ecefb750a3f5809ed5cb88662dac88fc453ad70
Reviewed-on: https://chromium-review.googlesource.com/750073
Commit-Ready: Rajat Jain <rajatja@chromium.org>
Tested-by: Rajat Jain <rajatja@chromium.org>
Reviewed-by: Rajat Jain <rajatja@chromium.org>
Reviewed-by: Todd Broch <tbroch@chromium.org>
(cherry picked from commit 629db59a0f14b0da36a2beb2ff075812370805d4)
Reviewed-on: https://chromium-review.googlesource.com/1641702
Commit-Ready: Rajat Jain <rajatja@google.com>
Tested-by: Rajat Jain <rajatja@google.com>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Sukumar Ghorai <sukumar.ghorai@intel.com>
Reviewed-by: Yoni Shavit <yshavit@chromium.org>