fingerprint: fpc: Use TRIGGER_HIGH for FPC interrupt
By using TRIGGER_RISING, we are risking to fire the interrupt handler way too many
times, even when we don't really need it to. Actually, we want it to fire when the
interrupt is HIGH, because our actions will be taken ONLY in this case. It would be
useless to read the GPIO state in the handler when the IRQ is rising when we can
just configure it as trigger high.

Change-Id: I44829f7aabacb29a11eec8e125bfba095c85e9ff
[ Tashar02: Port to fpc1020 ]
Signed-off-by: Tashfin Shakeer Rhythm <tashfinshakeerrhythm@gmail.com>