xiaomi: fpc: Fix and optimize regulators handling
It's useless to request and deregister a regulator everytime we bring one up/down.
Also, it's expensive to search for a regulator through its name in a struct containing
a string to get a match.

Optimize the regulators handling by introducing an enum containing the regulator map.
Request the regulators only at probe time, differentiate between required and optional
rails and use devm functions to benefit from the auto put on device detach.

This change also fixes the kernel logspam because of optional regulators not found.

Change-Id: I82652a12733b5c536d6c16c462c1701a751d6397
[ Tashar02: Port to fpc1020 and inline fpc1020_get_regulators() ]
Co-authored-by: Tashfin Shakeer Rhythm <tashfinshakeerrhythm@gmail.com>
Signed-off-by: Tashfin Shakeer Rhythm <tashfinshakeerrhythm@gmail.com>