CHROMIUM: config: Set cpufreq perf gov to default for arm64
This is the equivalent of:

  echo 'CONFIG_CPU_FREQ_DEFAULT_GOV_PERFORMANCE=y' >>\
    chromeos/config/arm64/common.config
  ./chromeos/scripts/kernelconfig olddefconfig

...atop a config that was normalized before this patch.

In general we want the default cpufreq governor to be performance to
improve boot speed.  We'll switch to a more sane governor (like
schedutil) after bootup.  Some pointers that should help those curious
figure out how that happens, at least on one board:
  src/platform2/power_manager/tools/cpufreq_config
  overlays/baseboard-cheza/.../cpufreq.conf

For a previous discussion about this, see CL:376302

BUG=b:111527336
TEST=System still ends up at the right scheduler after boot

Change-Id: I8fce7745577aad584ecfad78254e688faa679417
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1355743
Reviewed-by: Matthias Kaehlcke <mka@chromium.org>