UPSTREAM: media: videobuf2-v4l2: refuse qbuf if queue uses requests or vv.
Check if the vb2 queue uses requests, and if so refuse to
add buffers that are not part of a request. Also check for
the reverse: a vb2 queue did not use requests, and an attempt
was made to queue a buffer to a request.

We might relax this in the future, but for now just return
-EPERM in that case.

BUG=chromium:902656
TEST=Checked that Cheza was working with request API enabled.

Signed-off-by: Hans Verkuil <hans.verkuil@cisco.com>
Reviewed-by: Mauro Carvalho Chehab <mchehab+samsung@kernel.org>
Signed-off-by: Mauro Carvalho Chehab <mchehab+samsung@kernel.org>
(cherry picked from commit 9ac8090fda77f072815c209c80fb230e89cda18c)
Signed-off-by: Alexandre Courbot <acourbot@chromium.org>

Change-Id: I782ddb38197e515e5dfffd26254ec9d317424e94
Reviewed-on: https://chromium-review.googlesource.com/c/1325779
Reviewed-by: Tomasz Figa <tfiga@chromium.org>
Tested-by: Alexandre Courbot <acourbot@chromium.org>