UPSTREAM: media: vb2: add init_buffer buffer op
We need to initialize the request_fd field in struct vb2_v4l2_buffer
to -1 instead of the default of 0. So we need to add a new op that
is called when struct vb2_v4l2_buffer is allocated.

BUG=chromium:902656
TEST=Checked that Cheza was working with request API enabled.

Signed-off-by: Hans Verkuil <hans.verkuil@cisco.com>
Reviewed-by: Mauro Carvalho Chehab <mchehab+samsung@kernel.org>
Signed-off-by: Mauro Carvalho Chehab <mchehab+samsung@kernel.org>
(cherry picked from commit 8e013700bc12806d80f31ebe360916987f0e03df)
Signed-off-by: Alexandre Courbot <acourbot@chromium.org>

Change-Id: I3efa9d4bf46574f0e6a05e55fd1e366bb2394aff
Reviewed-on: https://chromium-review.googlesource.com/c/1325772
Reviewed-by: Tomasz Figa <tfiga@chromium.org>
Tested-by: Alexandre Courbot <acourbot@chromium.org>