UPSTREAM: drm/amd/amdgpu: Fix debugfs error handling
The error handling is wrong and "ent" could be NULL we when dereference
it to get "ent->d_inode".

The thing is that normally debugfs_create_file() is not supposed to
require (or have) any error handling.  That function does return error
pointers if debugfs is turned off but we know it's enable here.  When
it's enabled, then it returns NULL on error.

So what I did was I stripped out all the error handling except around
the i_size_write().  I could have just used a NULL check instead of an
IS_ERR_OR_NULL() but I figured this was more clear because that way you
don't have to look at the surrounding code to see whether debugfs is
enabled or not.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>
Reviewed-by: Christian Knig <christian.koenig@amd.com>
Signed-off-by: Alex Deucher <alexander.deucher@amd.com>

(cherry picked from commit d344b21bf405eed05963627bfed6dd3df422623c)

BUG=None
TEST=Tested on cheza, build tested everything else

Signed-off-by: Sean Paul <seanpaul@chromium.org>
Change-Id: I59524627dd5168858a11f71c98ed0972fcf8f93d
Reviewed-on: https://chromium-review.googlesource.com/1342340
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>